bug fix, dashboard rendering undefined weatherData props
The weather data state is set after the first rendering of each of
the components. As a result, the dashboard component was conditionally
rendering data from the weatherData state that was undefined becuase
it was the first render, before the weatherData state is set. Conditonal
renders added to the dashboard component.