Vector layers for points on map and view center now change with areas
when a new area is clicked on from the location menu the center of the
map is upadated as well as the red point of where that area is.
The key was using a useRef to target both the element and another
useRef to target the map state. Anything tht need to change targets
the map state in the second ref