bug fix, map not rendering on intial load
the map was not rendering on the initial load after deleting the
climbing areas in local state because when the climbing areas
are loaded as defaults their coordinate values are strings.
This was not a problems when the coordinates were taken from
the currentCoordinate state as it was initial set the values
from the current location parsed as floats. Upon removing the
currentCoordinate state the default climbing area coordinates
were not being parsed into floats.