bug fix: svgs not compiling into dist folder
installed vitesvgr library which reads svg assets and turns them
into react components. changed getWeatherIcon function to conditionally
render a weather icon based on the weather code from the api. Had
some trouble getting the exported component to render until realizing
that it was rendering, but the height and width were not set