users now able to add point and go to that area after clicking
the challenge with this bug was the internals of open layers itself
when a user adds a point by clicking the map, pressing that area
would blow up the app. The reason was the logic for setting the
new location was within an event closure so even though
refs were passed in, the context of the state was the same as
on the initial render. Additionally there was no way to associate
the id the of the feature that was clicked to the area in the climbing
area state. To fix the same area id was put on the POINT FEATURE and
not the vector layer. Additionally a state change function was passed
into the event to set the id of the clicked feature. This was then
passed into a function within a useEffect outside of the event and the
correct area was filtered and set as the new location