bug fix, rock type defaults inaccurate upon initial page load
The rock data state was being initialized with a hard coded primary
rock class and kinds of rock object. This was ok when if the user is
loading the app for the first time but became problematic if the user
adds or deletes and area, closes the page and then reloads. To fix I
removed the hard coded intialized state and added the rock types state
to the dependcy array that gets the data from the rock api. Now when
the app loads and rock type is set it triggers the useEffect to set
the rock data.

Additionally I replaced the conditional render of the table component
with logic inside of the table compoenent that checks what the button
title state is. If it is Wet Rock then a display none class is added
instead of flex. This problem was caused by the above fix becuase when
the component was not rendered on a button change, the state in both
rock types and rock data was erased.