rock parsing logic simplified
the js responsible for parsing the current location rock data
was streamlined. Instead of parsing small amounts of data and then
aggregating data is handled in one helper. Rock types is now
set as a map so that each time a new loction is switched to all of the
rock type data does not have to be iterated through.